shopImageTrainBackend/requirements.txt
ultralytics>=8.0.0
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.5.0
Pillow>=8.3.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
PyYAML>=6.0
tqdm>=4.62.0
flask>=2.0.0
flask-cors>=3.0.0
werkzeug>=2.0.0
python-multipart>=0.0.5
pymysql>=1.0.2
sqlalchemy>=1.4.0
minio>=7.1.0
flask-socketio>=5.0.0
python-socketio>=5.0.0